refactor(decompose): split packages/libraries/graph-tools/scripts/L3-BF-191-ascii-parser.ts into folder
